\name{GeneTS-internal}
\alias{dominant.freqs.single}
\alias{periodogram.spec.single}
\alias{periodogram.spec}
\alias{periodogram.freq}
\alias{fisher.g.test.single}
\alias{is.constant.single}
\alias{bag.fun}
\alias{myrmvnorm}

\title{Internal GeneTS Functions}
\description{
  Internal GeneTS functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
